Add support for date related temporal accessors
